Preliminary changes to allow modelGroup::sequence to support maxOccurs > 1.
Massive code cleanup of wsdl-types-procs and <ws>proc

Children of sequence elements are protected, they are no longer
 global types. Only sequence elements and constructed simpleTypes
 are global.

Result of previous cleanup is that there is less chaining of types.
 simpleTypes are not imported into a namespace just to rename them,
 but are used in the children of sequences (element type).

modelGroup::sequence::new now can include complexTypes

All sequence specifications are not identical. Previously we had:
 {TypeName baseType minOccurs maxOccurs {facetList}} for tWSDL
and
 {TypeName:baseType {facetList}} for <ws>type and <ws>element
and
 {{Typename:base {defaultValue}}...} for <ws>proc arguments

Now we have:
 {TypeName:baseType {facetList}} for every case.

This means that changes are necessary to configuration files:

<ws>proc ::ns::myproc {
 Arg1 
 {Arg2 mydefault}
} { }... 

is now changed to:

<ws>proc ::ns::myproc {
 Arg1 
 {Arg2  {minOccurs 0 default "mydefault"}} 
} { }...

See examples under packages/wsapi/www/*/index/tcl