Additional minor bugs fixed.
Most important: the canonical value for restricted decimals
which have fractionDigits = 0, now removes the decimal point.

However, note that any datatype which is so restricted should
be derived from an integer type, due to the fact that a value
with a trailing decimal point, but no trailing digits, will still
validate as an integer if the base type is decimal.